@@ -2,11 +2,11 @@
   <PropertyGroup>
     <BuildDocFx Condition=" '$(Configuration)'=='Debug' ">false</BuildDocFx>  <!-- Set this value to 'true' in order to build the documentation. -->
     <InvariantGlobalization>true</InvariantGlobalization>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <Authors>Fernando Porrino Serrano</Authors>
     <Product>AutoCheck.Core</Product>
     <Copyright>Copyright © 2021</Copyright>
-    <VersionPrefix>2.11.1</VersionPrefix>
+    <VersionPrefix>2.12.0</VersionPrefix>
     <VersionSuffix>stable</VersionSuffix>
     <AssemblyVersion>$(VersionPrefix)</AssemblyVersion> 
     <AssemblyFileVersion>$(AssemblyVersion)</AssemblyFileVersion>
