@@ -35,11 +35,11 @@ namespace AutoCheck.Test.Connectors
         }
 
         [Test]
-        [TestCase("correct.atom", OS.GNU, "localhost", "usuario", "usuario")]
+        [TestCase("correct.atom", OS.GNU, "localhost", "autocheck", "autocheck")]
         public void Constructor_Remote_DoesNotThrow(string file, OS remoteOS, string host, string username, string password)
         {     
             //Note: the source code for local and remote mode are exactly the same, just need to test that the remote file is being downloaded from remote and parsed. 
-            Assert.DoesNotThrow(() => new AutoCheck.Core.Connectors.Atom(remoteOS, host, username, password, LocalPathToWsl(GetSampleFile(file))));
+            Assert.DoesNotThrow(() => new AutoCheck.Core.Connectors.Atom(remoteOS, host, username, password, LocalPathToRemote(GetSampleFile(file), username)));
         }
 
         [Test]
