@@ -45,25 +45,25 @@ namespace AutoCheck.Test.Connectors
         }
    
         [Test]
-        [TestCase("", OS.GNU, "localhost", "usuario", "usuario")]
+        [TestCase("", OS.GNU, "localhost", "autocheck", "autocheck")]
         public void Constructor_Remote_Throws_ArgumentNullException(string file, OS remoteOS, string host, string username, string password)
         {     
             Assert.Throws<ArgumentNullException>(() => new AutoCheck.Core.Connectors.Html(remoteOS, host, username, password, file));
         }
 
         [Test]
-        [TestCase(_FAKE, OS.GNU, "localhost", "usuario", "usuario")]
+        [TestCase(_FAKE, OS.GNU, "localhost", "autocheck", "autocheck")]
         public void Constructor_Remote_Throws_FileNotFoundException(string file, OS remoteOS, string host, string username, string password)
         {     
             Assert.Throws<FileNotFoundException>(() => new AutoCheck.Core.Connectors.Html(remoteOS, host, username, password, file));
         }
 
         [Test]
-        [TestCase("correct.html", OS.GNU, "localhost", "usuario", "usuario")]
+        [TestCase("correct.html", OS.GNU, "localhost", "autocheck", "autocheck")]
         public void Constructor_DoesNotThrow(string file, OS remoteOS, string host, string username, string password)
         {           
             //Note: the source code for local and remote mode are exactly the same, just need to test that the remote file is being downloaded from remote and parsed.
-            Assert.DoesNotThrow(() => new AutoCheck.Core.Connectors.Html(remoteOS, host, username, password, LocalPathToWsl(GetSampleFile(file))));            
+            Assert.DoesNotThrow(() => new AutoCheck.Core.Connectors.Html(remoteOS, host, username, password, LocalPathToRemote(GetSampleFile(file), username)));            
         }
 
         [Test]
