@@ -39,14 +39,14 @@ namespace AutoCheck.Test.Connectors
             //Remote
             const OS remoteOS = OS.GNU;
             const string host = "localhost";
-            const string username = "usuario";
-            const string password = "usuario";
+            const string username = "autocheck";
+            const string password = "autocheck";
 
             Assert.Throws<ArgumentNullException>(() => new AutoCheck.Core.Connectors.Rss(remoteOS, host, username, password, string.Empty));
             Assert.Throws<FileNotFoundException>(() => new AutoCheck.Core.Connectors.Rss(remoteOS, host, username, password, _FAKE));
 
             //Note: the source code for local and remote mode are exactly the same, just need to test that the remote file is being downloaded from remote and parsed.
-            var file = LocalPathToWsl(GetSampleFile("correct.rss"));
+            var file = LocalPathToRemote(GetSampleFile("correct.rss"), username);
             Assert.DoesNotThrow(() => new AutoCheck.Core.Connectors.Rss(OS.GNU, host, username, password, file));
         }
 
